full program

+++ Coin Flipper Activity/Game.h
@@ -12,7 +12,7 @@ private:
 	Coin* coins;
 	int numberOfCoins;
 	double targetBalance;
-	int totalScore;
+	double totalScore;
 	int rounds;
